Switch to k8s test app
