+++ pyproject.toml
@@ -60,18 +60,12 @@ pytest-operator = "0.29.0"
 juju = "2.9.44.0 || 3.2.0.1"  # renovate libjuju
 psycopg2 = {version = "^2.9.5", extras = ["binary"]}
 
-[tool.poetry.group.ha_charm]
-optional = true
-
-[tool.poetry.group.ha_charm.dependencies]
-ops = "2.5.1"
-tenacity = "8.2.3"
-
 [tool.poetry.group.relation_charm]
 optional = true
 
 [tool.poetry.group.relation_charm.dependencies]
 ops = "2.5.1"
+tenacity = "8.2.3"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
