--- tests/integration/ha_tests/application-charm/metadata.yaml
@@ -1,17 +1,0 @@
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-name: application
-description: |
-  Data platform libs application charm used in high availability integration tests.
-summary: |
-  Data platform libs application meant to be used
-  only for testing high availability of the PostgreSQL charm.
-
-requires:
-  database:
-    interface: postgresql_client
-    limit: 1
-
-peers:
-  application-peers:
-    interface: application-peers
