--- tests/integration/new_relations/application-charm/requirements.txt
@@ -1,3 +1,0 @@
-ops==2.5.1 ; python_full_version >= "3.10.6" and python_full_version < "4.0.0"
-pyyaml==6.0.1 ; python_full_version >= "3.10.6" and python_full_version < "4.0.0"
-websocket-client==1.6.2 ; python_full_version >= "3.10.6" and python_full_version < "4.0.0"
