--- tests/integration/new_relations/conftest.py
@@ -1,21 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-import pytest
-from pytest_operator.plugin import OpsTest
-
-
-@pytest.fixture(scope="module")
-async def application_charm(ops_test: OpsTest):
-    """Build the application charm."""
-    charm_path = "tests/integration/new_relations/application-charm"
-    charm = await ops_test.build_charm(charm_path)
-    return charm
-
-
-@pytest.fixture(scope="module")
-async def database_charm(ops_test: OpsTest):
-    """Build the database charm."""
-    charm = await ops_test.build_charm(".")
-    return charm
