+++ tox.ini
@@ -198,5 +198,4 @@ description = Install, lock and export poetry dependencies
 commands =
     poetry lock
     poetry export -f requirements.txt -o requirements.txt
-    poetry export --only ha_charm -f requirements.txt -o tests/integration/ha_tests/application-charm/requirements.txt
-    poetry export --only relation_charm -f requirements.txt -o tests/integration/new_relations/application-charm/requirements.txt
+    poetry export --only relation_charm -f requirements.txt -o tests/integration/postgresql-test-app/requirements.txt
